+++ ngClient/_system/js/app/controllers/account/AddPasswordController.js
@@ -3,16 +3,16 @@
 
     var controllerId = 'AddPasswordController';
     angular.module('main')
-        .controller(controllerId, ['userFactory', '$location', 'logger', AddPasswordController]);
+        .controller(controllerId, ['dataContext', '$location', 'logger', AddPasswordController]);
 
-    function AddPasswordController(userFactory, $location, logger) {
+    function AddPasswordController(dataContext, $location, logger) {
         logger = logger.forSource(controllerId);
 
         var vm = this;
         vm.addPassword = addPassword;
 
         function addPassword() {
-            userFactory.addPassword(vm)
+            dataContext.addPassword(vm)
                 .success(function () {
                     $location.url('/');
                     logger.logSuccess('Your password has been set!', null, true);