+++ ngClient/_system/js/app/controllers/account/ChangeEmailController.js
@@ -3,9 +3,9 @@
 
     var controllerId = 'ChangeEmailController';
     angular.module('main')
-        .controller(controllerId, ['userFactory', '$location', 'logger', ChangeEmailController]);
+        .controller(controllerId, ['dataContext', '$location', 'logger', ChangeEmailController]);
 
-    function ChangeEmailController(userFactory, $location, logger) {
+    function ChangeEmailController(dataContext, $location, logger) {
         logger = logger.forSource(controllerId);
 
         var vm = this;
@@ -18,16 +18,7 @@
 
             // Generate test data if localhost
             if ($location.host() === 'localhost') {
-                var now = new Date();
-                var year = now.getFullYear();
-                var month = now.getMonth() + 1;
-                var day = now.getDate();
-                var hour = now.getHours();
-                var minute = now.getMinutes();
-                var second = now.getSeconds();
-                var email = 'local_' + year + month + day + '_' + hour + minute + second + '@forcrowd.org';
-
-                vm.email = email;
+                vm.email = dataContext.getUniqueUserEmail();
             }
         }
 
@@ -35,7 +26,7 @@
 
             vm.isChangeEmailDisabled = true;
 
-            userFactory.changeEmail(vm)
+            dataContext.changeEmail(vm)
                 .success(function () {
                     $location.url('/_system/account/confirmEmail');
                 })