+++ ngClient/_system/js/app/controllers/account/LoginController.js
@@ -3,17 +3,18 @@
 
     var controllerId = 'LoginController';
     angular.module('main')
-        .controller(controllerId, ['userFactory', '$location', 'locationHistory', 'serviceAppUrl', 'logger', LoginController]);
+        .controller(controllerId, ['dataContext', 'locationHistory', 'serviceAppUrl', 'logger', '$location', '$rootScope', '$scope', LoginController]);
 
-    function LoginController(userFactory, $location, locationHistory, serviceAppUrl, logger) {
+    function LoginController(dataContext, locationHistory, serviceAppUrl, logger, $location, $rootScope, $scope) {
         logger = logger.forSource(controllerId);
 
         var vm = this;
         vm.email = '';
         vm.login = login;
-        vm.getExternalLoginUrl = getExternalLoginUrl;
+        vm.navigateToResetPassword = navigateToResetPassword;
         vm.password = '';
-        vm.rememberMe = false;
+        vm.rememberMe = true;
+        vm.showHeader = typeof $scope.showHeader !== 'undefined' ? $scope.showHeader : true;
 
         _init();
 
@@ -29,15 +30,21 @@
             login();
         }
 
+        function getReturnUrl() {
+            // If login pages called after a result from server, it will have "clientReturnUrl" param, which will have a higher priority than locationHistory
+            var clientReturnUrl = $location.search().clientReturnUrl;
+            return typeof clientReturnUrl !== 'undefined' ? clientReturnUrl : locationHistory.previousItem().url();
+        }
+
         function login() {
 
-            // External (temp token) login
-            var tempToken = $location.search().tempToken;
-            if (typeof tempToken !== 'undefined') {
-                userFactory.getToken('', '', vm.rememberMe, tempToken).then(success).catch(failedExternal);
+            // External (single use token) login
+            var singleUseToken = $location.search().token;
+            if (typeof singleUseToken !== 'undefined') {
+                dataContext.login('', '', vm.rememberMe, singleUseToken).then(success).catch(failedExternal);
             } else { // Internal login
                 if (vm.email !== '' && vm.password !== '') {
-                    userFactory.getToken(vm.email, vm.password, vm.rememberMe).then(success);
+                    dataContext.login(vm.email, vm.password, vm.rememberMe).then(success);
                 }
             }
 
@@ -51,14 +58,9 @@
             }
         }
 
-        function getExternalLoginUrl(provider) {
-            return serviceAppUrl + '/api/Account/ExternalLogin?provider=' + provider + '&clientReturnUrl=' + getReturnUrl();
-        }
-
-        function getReturnUrl() {
-            // If login pages called after a result from server, it will have "clientReturnUrl" param, which will have a higher priority than locationHistory
-            var clientReturnUrl = $location.search().clientReturnUrl;
-            return typeof clientReturnUrl !== 'undefined' ? clientReturnUrl : locationHistory.previousItem().url();
+        function navigateToResetPassword() {
+            $rootScope.$broadcast('LoginController_redirected');
+            $location.path('/_system/account/resetPassword');
         }
     }
 })();